--- src/main/java/com/cristinaj/loganalyzer/LogEntryRepository.java
@@ -1,9 +1,0 @@
-package com.cristinaj.loganalyzer;
-
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface LogEntryRepository extends JpaRepository<LogEntryEntity, Long> {
-    Page<LogEntryEntity> findByLevel(LogLevel level, Pageable pageable);
-}
