+++ src/main/java/com/cristinaj/loganalyzer/domain/LogLevel.java
@@ -1,4 +1,4 @@
-package com.cristinaj.loganalyzer;
+package com.cristinaj.loganalyzer.domain;
 
 public enum LogLevel {
     TRACE, DEBUG, INFO, WARN, ERROR
