+++ src/main/java/com/cristinaj/loganalyzer/web/ApiError.java
@@ -1,3 +1,3 @@
-package com.cristinaj.loganalyzer;
+package com.cristinaj.loganalyzer.web;
 
 public record ApiError(String error, String message) {}
